refactor: separate meta_impl from meta/lib.rs
