+++ cluster/src/config.rs
@@ -1,6 +1,6 @@
 // Copyright 2022 CeresDB Project Authors. Licensed under Apache-2.0.
 
-use meta_client_v2::MetaClientConfig;
+use meta_client_v2::meta_impl::MetaClientConfig;
 use serde_derive::Deserialize;
 
 #[derive(Default, Clone, Deserialize, Debug)]
