+++ cluster/src/lib.rs
@@ -4,11 +4,11 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 use common_util::define_result;
-pub use meta_client_v2::{
+pub use meta_client_v2::types::{
     AllocSchemaIdRequest, AllocSchemaIdResponse, AllocTableIdRequest, AllocTableIdResponse,
     DropTableRequest, DropTableResponse, GetTablesRequest,
 };
-use meta_client_v2::{ShardId, TableId};
+use meta_client_v2::types::{ShardId, TableId};
 use snafu::{Backtrace, Snafu};
 
 pub mod cluster_impl;
