+++ meta_client_v2/src/lib.rs
@@ -1,42 +1,18 @@
 // Copyright 2022 CeresDB Project Authors. Licensed under Apache-2.0.
 
-use std::{
-    sync::{Arc, Mutex, Weak},
-    time::Duration,
-};
+use std::sync::Arc;
 
 use async_trait::async_trait;
-use ceresdbproto_deps::ceresdbproto::{
-    meta_service::{
-        AllocSchemaIdRequest as PbAllocSchemaIdRequest,
-        AllocTableIdRequest as PbAllocTableIdRequest, DropTableRequest as PbDropTableRequest,
-        GetTablesRequest as PbGetTablesRequest, NodeHeartbeatRequest as PbNodeHeartbeatRequest,
-        NodeHeartbeatResponse as PbNodeHeartbeatResponse,
-    },
-    meta_service_grpc::CeresmetaRpcServiceClient,
+use common_util::define_result;
+use snafu::{Backtrace, Snafu};
+use types::{
+    ActionCmd, AllocSchemaIdRequest, AllocSchemaIdResponse, AllocTableIdRequest,
+    AllocTableIdResponse, DropTableRequest, DropTableResponse, GetTablesRequest, GetTablesResponse,
+    ResponseHeader, ShardInfo,
 };
-use common_util::{
-    config::ReadableDuration,
-    define_result,
-    runtime::{JoinHandle, Runtime},
-};
-use futures::{SinkExt, TryStreamExt};
-use grpcio::{
-    CallOption, ChannelBuilder, ClientDuplexReceiver, ClientDuplexSender, Environment, WriteFlags,
-};
-use log::{debug, error, info, warn};
-use serde_derive::Deserialize;
-use snafu::{Backtrace, OptionExt, ResultExt, Snafu};
-use tokio::{
-    sync::{
-        mpsc::{self, Receiver, Sender},
-        RwLock,
-    },
-    time,
-};
-pub use types::*;
 
-mod types;
+pub mod meta_impl;
+pub mod types;
 
 #[derive(Debug, Snafu)]
 #[snafu(visibility = "pub")]
@@ -73,16 +49,6 @@ pub enum Error {
         source: Box<dyn std::error::Error + Send + Sync>,
     },
 
-    #[snafu(display(
-        "Failed to notify action cmd, action cmd:{:?}, err:{}",
-        action_cmd,
-        source
-    ))]
-    FailNotifyActionCmd {
-        action_cmd: ActionCmd,
-        source: Box<dyn std::error::Error + Send + Sync>,
-    },
-
     #[snafu(display("Failed to alloc schema id, err:{}", source))]
     FailAllocSchemaId {
         source: Box<dyn std::error::Error + Send + Sync>,
@@ -103,19 +69,19 @@ pub enum Error {
         source: Box<dyn std::error::Error + Send + Sync>,
     },
 
-    #[snafu(display("Meta error, resp header:{:?}.\nBacktrace:\n{}", header, backtrace))]
-    Meta {
+    #[snafu(display("Meta rpc error, resp header:{:?}.\nBacktrace:\n{}", header, backtrace))]
+    MetaRpc {
         header: ResponseHeader,
         backtrace: Backtrace,
     },
 
     #[snafu(display(
-        "Handler event failed, handler:{}, event:{:?}, err:{}",
+        "Handle event failed, handler:{}, event:{:?}, err:{}",
         name,
         event,
         source
     ))]
-    FailHandlerEvent {
+    FailHandleEvent {
         name: String,
         event: ActionCmd,
         source: Box<dyn std::error::Error + Send + Sync>,
@@ -136,7 +102,8 @@ pub trait EventHandler {
     ) -> std::result::Result<(), Box<dyn std::error::Error + Send + Sync>>;
 }
 
-/// Meta client abstraction
+/// MetaClient is the abstraction of client used to communicate with CeresMeta
+/// cluster.
 #[async_trait]
 pub trait MetaClient {
     /// Start the meta client and the events will occur afterwards.
@@ -158,470 +125,4 @@ pub trait MetaClient {
     async fn get_tables(&self, req: GetTablesRequest) -> Result<GetTablesResponse>;
 
     async fn send_heartbeat(&self, req: Vec<ShardInfo>) -> Result<()>;
-}
-
-#[derive(Debug, Deserialize, Clone)]
-#[serde(default)]
-pub struct MetaClientConfig {
-    pub cluster_name: String,
-    pub meta_addr: String,
-    pub lease: ReadableDuration,
-    pub timeout: ReadableDuration,
-    pub cq_count: usize,
-}
-
-impl Default for MetaClientConfig {
-    fn default() -> Self {
-        Self {
-            cluster_name: String::new(),
-            meta_addr: "http://127.0.0.1:8080".to_string(),
-            lease: ReadableDuration::secs(10),
-            timeout: ReadableDuration::secs(5),
-            cq_count: 8,
-        }
-    }
-}
-
-struct NodeHeartbeatChannel {
-    heartbeat_sender: ClientDuplexSender<PbNodeHeartbeatRequest>,
-    action_cmd_receiver: Option<ClientDuplexReceiver<PbNodeHeartbeatResponse>>,
-}
-
-struct GrpcClient {
-    client: CeresmetaRpcServiceClient,
-    heartbeat_channel: RwLock<NodeHeartbeatChannel>,
-}
-
-// The handler may contain the Self instance which will cause circular
-// dependency.
-type EventHandlers = Vec<Weak<dyn EventHandler + Send + Sync>>;
-
-struct MetaClientImplInner {
-    meta_config: MetaClientConfig,
-    node_meta_info: NodeMetaInfo,
-
-    grpc_client: RwLock<Option<GrpcClient>>,
-    handlers: RwLock<EventHandlers>,
-}
-
-impl MetaClientImplInner {
-    fn new(meta_config: MetaClientConfig, node_meta_info: NodeMetaInfo) -> Result<Self> {
-        let client = Self {
-            meta_config,
-            node_meta_info,
-            grpc_client: RwLock::new(None),
-            handlers: RwLock::new(EventHandlers::new()),
-        };
-
-        Ok(client)
-    }
-
-    fn request_header(&self) -> RequestHeader {
-        RequestHeader {
-            node: self.node_meta_info.node.to_string(),
-            cluster_name: self.meta_config.cluster_name.clone(),
-        }
-    }
-
-    fn node_meta_info(&self) -> NodeMetaInfo {
-        self.node_meta_info.clone()
-    }
-
-    fn get_cluster_name(&self) -> &str {
-        self.meta_config.cluster_name.as_str()
-    }
-
-    fn connect_grpc_client(&self) -> Result<GrpcClient> {
-        let client = self.build_rpc_client();
-        let (sender, receiver) = client
-            .node_heartbeat_opt(CallOption::default())
-            .context(InitHeartBeatStream)?;
-        Ok(GrpcClient {
-            client,
-            heartbeat_channel: RwLock::new(NodeHeartbeatChannel {
-                heartbeat_sender: sender,
-                action_cmd_receiver: Some(receiver),
-            }),
-        })
-    }
-
-    async fn reconnect_heartbeat_channel(&self) {
-        info!("gRPC reconnect begin");
-
-        loop {
-            match self.connect_grpc_client() {
-                Ok(client) => {
-                    *self.grpc_client.write().await = Some(client);
-
-                    info!("gRPC reconnect succeeds");
-                    return;
-                }
-                Err(e) => {
-                    error!("gRPC reconnect failed, err:{}", e);
-
-                    time::sleep(self.error_wait_lease()).await;
-                }
-            }
-        }
-    }
-
-    fn build_rpc_client(&self) -> CeresmetaRpcServiceClient {
-        let cb = ChannelBuilder::new(Arc::new(Environment::new(self.meta_config.cq_count)));
-        CeresmetaRpcServiceClient::new(cb.connect(&self.meta_config.meta_addr))
-    }
-
-    // TODO: Store the value in field
-    fn error_wait_lease(&self) -> Duration {
-        Duration::from_secs(self.meta_config.lease.as_secs() / 2)
-    }
-
-    async fn fetch_action_cmd_loop(&self, mut stop_rx: Receiver<()>) {
-        info!("Begin fetching action cmd loop");
-        loop {
-            debug!("Begin fetching action cmd once");
-            let receiver = match &*self.grpc_client.read().await {
-                Some(client) => client
-                    .heartbeat_channel
-                    .write()
-                    .await
-                    .action_cmd_receiver
-                    .take(),
-                None => {
-                    warn!("gRPC client is not inited when starting fetch action cmd");
-                    None
-                }
-            };
-
-            if let Some(v) = receiver {
-                match self.fetch_action_cmd(v).await {
-                    Ok(()) => {
-                        info!(
-                            "Fetch cluster action cmd finished, cluster:{}",
-                            self.get_cluster_name()
-                        );
-                    }
-                    Err(err) => {
-                        error!(
-                            "Failed to get action cmd, cluster:{}, err:{}",
-                            self.get_cluster_name(),
-                            err
-                        );
-                        self.reconnect_heartbeat_channel().await;
-                    }
-                }
-            } else {
-                warn!("Skip action command fetch because no receiver found");
-            }
-
-            if time::timeout(self.error_wait_lease(), stop_rx.recv())
-                .await
-                .is_ok()
-            {
-                warn!("Receive exit message, exit the fetch_action_cmd loop");
-                break;
-            }
-        }
-    }
-
-    async fn fetch_action_cmd(
-        &self,
-        mut receiver: ClientDuplexReceiver<PbNodeHeartbeatResponse>,
-    ) -> Result<()> {
-        info!("Start fetch action cmd loop");
-        while let Some(resp) = receiver.try_next().await.context(FetchActionCmd)? {
-            info!(
-                "Fetch action cmd from meta, cluster:{}, action_cmd:{:?}",
-                self.get_cluster_name(),
-                resp,
-            );
-
-            let resp = NodeHeartbeatResponse::from(resp);
-            if let Err(e) = check_response_header(&resp.header) {
-                error!("Fetch action cmd failed, err:{}", e);
-                continue;
-            }
-            let event = match resp.action_cmd {
-                Some(action_cmd) => action_cmd,
-                None => {
-                    warn!("Fetch action cmd is empty, resp:{:?}", resp);
-                    continue;
-                }
-            };
-
-            if let Err(e) = self.handle_event(&event).await {
-                error!("Handler fail to handle event:{:?}, err:{}", event, e);
-            }
-        }
-
-        Ok(())
-    }
-
-    pub async fn handle_event(&self, event: &ActionCmd) -> Result<()> {
-        let handlers = self.handlers.read().await;
-        for handler in handlers.iter() {
-            if let Some(handler) = handler.upgrade() {
-                handler.handle(event).await.context(FailHandlerEvent {
-                    name: handler.name().to_string(),
-                    event: event.clone(),
-                })?;
-            }
-        }
-        Ok(())
-    }
-}
-
-/// Default meta client impl, will interact with a remote meta node.
-pub struct MetaClientImpl {
-    inner: Arc<MetaClientImplInner>,
-    runtime: Arc<Runtime>,
-
-    eventloop_handle: Mutex<Option<JoinHandle<()>>>,
-    stop_eventloop_tx: Mutex<Option<Sender<()>>>,
-}
-
-impl MetaClientImpl {
-    pub fn new(
-        config: MetaClientConfig,
-        node_meta_info: NodeMetaInfo,
-        runtime: Arc<Runtime>,
-    ) -> Result<Self> {
-        Ok(Self {
-            inner: Arc::new(MetaClientImplInner::new(config, node_meta_info)?),
-            runtime,
-            eventloop_handle: Mutex::new(None),
-            stop_eventloop_tx: Mutex::new(None),
-        })
-    }
-}
-
-#[async_trait]
-impl MetaClient for MetaClientImpl {
-    async fn start(&self) -> Result<()> {
-        info!(
-            "Meta client is starting with config:{:?}",
-            self.inner.meta_config
-        );
-
-        self.inner.reconnect_heartbeat_channel().await;
-
-        let inner = self.inner.clone();
-        let (tx, rx) = mpsc::channel(1);
-        let eventloop_handle = self.runtime.spawn(async move {
-            inner.fetch_action_cmd_loop(rx).await;
-        });
-
-        *self.eventloop_handle.lock().unwrap() = Some(eventloop_handle);
-        *self.stop_eventloop_tx.lock().unwrap() = Some(tx);
-
-        info!("Meta client has started");
-        Ok(())
-    }
-
-    async fn stop(&self) -> Result<()> {
-        info!("Meta client is stopping");
-
-        {
-            let tx = self.stop_eventloop_tx.lock().unwrap().take();
-            if let Some(tx) = tx {
-                let _ = tx.send(()).await;
-            }
-        }
-
-        {
-            let handle = self.eventloop_handle.lock().unwrap().take();
-            if let Some(handle) = handle {
-                let _ = handle.await;
-            }
-        }
-
-        info!("Meta client has stopped");
-        Ok(())
-    }
-
-    async fn register_event_handler(&self, handler: EventHandlerRef) -> Result<()> {
-        let mut handlers = self.inner.handlers.write().await;
-        handlers.push(Arc::downgrade(&handler));
-
-        Ok(())
-    }
-
-    async fn alloc_schema_id(&self, req: AllocSchemaIdRequest) -> Result<AllocSchemaIdResponse> {
-        let grpc_client_guard = self.inner.grpc_client.read().await;
-        let grpc_client = grpc_client_guard.as_ref().context(FailGetGrpcClient)?;
-
-        let mut pb_req = PbAllocSchemaIdRequest::from(req);
-        pb_req.set_header(self.inner.request_header().into());
-        let pb_resp = grpc_client
-            .client
-            .alloc_schema_id_async_opt(&pb_req, CallOption::default())
-            .map_err(|e| Box::new(e) as _)
-            .context(FailAllocSchemaId)?
-            .await
-            .map_err(|e| Box::new(e) as _)
-            .context(FailAllocSchemaId)?;
-
-        debug!(
-            "Meta client alloc schema id, req:{:?}, resp:{:?}",
-            pb_req, pb_resp
-        );
-
-        let resp = AllocSchemaIdResponse::from(pb_resp);
-
-        check_response_header(&resp.header)?;
-        Ok(resp)
-    }
-
-    async fn alloc_table_id(&self, req: AllocTableIdRequest) -> Result<AllocTableIdResponse> {
-        let grpc_client_guard = self.inner.grpc_client.read().await;
-        let grpc_client = grpc_client_guard.as_ref().context(FailGetGrpcClient)?;
-
-        let mut pb_req = PbAllocTableIdRequest::from(req);
-        pb_req.set_header(self.inner.request_header().into());
-        let pb_resp = grpc_client
-            .client
-            .alloc_table_id_async_opt(&pb_req, CallOption::default())
-            .map_err(|e| Box::new(e) as _)
-            .context(FailAllocTableId)?
-            .await
-            .map_err(|e| Box::new(e) as _)
-            .context(FailAllocTableId)?;
-
-        debug!(
-            "Meta client alloc table id, req:{:?}, resp:{:?}",
-            pb_req, pb_resp
-        );
-
-        let resp = AllocTableIdResponse::from(pb_resp);
-        check_response_header(&resp.header)?;
-
-        let add_table_cmd = ActionCmd::AddTableCmd(AddTableCmd {
-            schema_name: resp.schema_name.clone(),
-            name: resp.name.clone(),
-            shard_id: resp.shard_id,
-            schema_id: resp.schema_id,
-            id: resp.id,
-        });
-        self.inner.handle_event(&add_table_cmd).await?;
-
-        Ok(resp)
-    }
-
-    async fn drop_table(&self, req: DropTableRequest) -> Result<DropTableResponse> {
-        let grpc_client_guard = self.inner.grpc_client.read().await;
-        let grpc_client = grpc_client_guard.as_ref().context(FailGetGrpcClient)?;
-
-        let mut pb_req = PbDropTableRequest::from(req.clone());
-        pb_req.set_header(self.inner.request_header().into());
-        let pb_resp = grpc_client
-            .client
-            .drop_table_async_opt(&pb_req, CallOption::default())
-            .map_err(|e| Box::new(e) as _)
-            .context(FailDropTable)?
-            .await
-            .map_err(|e| Box::new(e) as _)
-            .context(FailDropTable)?;
-
-        debug!(
-            "Meta client drop table, req:{:?}, resp:{:?}",
-            pb_req, pb_resp
-        );
-
-        let resp = DropTableResponse::from(pb_resp);
-        check_response_header(&resp.header)?;
-
-        let drop_table_cmd = ActionCmd::DropTableCmd(DropTableCmd {
-            schema_name: req.schema_name.clone(),
-            name: req.name.clone(),
-        });
-        self.inner.handle_event(&drop_table_cmd).await?;
-
-        Ok(resp)
-    }
-
-    async fn get_tables(&self, req: GetTablesRequest) -> Result<GetTablesResponse> {
-        let grpc_client_guard = self.inner.grpc_client.read().await;
-        let grpc_client = grpc_client_guard.as_ref().context(FailGetGrpcClient)?;
-
-        let mut pb_req = PbGetTablesRequest::from(req);
-        pb_req.set_header(self.inner.request_header().into());
-
-        let pb_resp = grpc_client
-            .client
-            .get_tables_async_opt(&pb_req, CallOption::default())
-            .map_err(|e| Box::new(e) as _)
-            .context(FailGetTables)?
-            .await
-            .map_err(|e| Box::new(e) as _)
-            .context(FailGetTables)?;
-
-        debug!(
-            "Meta client get tables, req:{:?}, resp:{:?}",
-            pb_req, pb_resp
-        );
-
-        let resp = GetTablesResponse::from(pb_resp);
-        check_response_header(&resp.header)?;
-
-        Ok(resp)
-    }
-
-    async fn send_heartbeat(&self, shards_info: Vec<ShardInfo>) -> Result<()> {
-        info!(
-            "Meta client send heartbeat, cluster:{}, shards_info:{:?}",
-            self.inner.get_cluster_name(),
-            shards_info
-        );
-
-        let grpc_client_guard = self.inner.grpc_client.read().await;
-        let grpc_client = grpc_client_guard.as_ref().context(FailGetGrpcClient)?;
-
-        let mut pb_request = PbNodeHeartbeatRequest::new();
-        pb_request.set_header(self.inner.request_header().into());
-        let node_info = NodeInfo {
-            node_meta_info: self.inner.node_meta_info(),
-            shards_info,
-        };
-        pb_request.set_info(node_info.into());
-
-        info!("Meta client send heartbeat req:{:?}", pb_request);
-
-        let send_res = grpc_client
-            .heartbeat_channel
-            .write()
-            .await
-            .heartbeat_sender
-            .send((pb_request, WriteFlags::default()))
-            .await
-            .map_err(|e| Box::new(e) as _)
-            .context(FailSendHeartbeat {
-                cluster: self.inner.get_cluster_name(),
-            });
-
-        // FIXME: reconnect the heartbeat channel iff specific errors occur.
-        if send_res.is_err() {
-            self.inner.reconnect_heartbeat_channel().await;
-        }
-        send_res
-    }
-}
-
-fn check_response_header(header: &ResponseHeader) -> Result<()> {
-    if header.is_success() {
-        Ok(())
-    } else {
-        Meta {
-            header: header.clone(),
-        }
-        .fail()
-    }
-}
-
-/// Create a meta client with given `config`.
-pub fn build_meta_client(
-    config: MetaClientConfig,
-    node_meta_info: NodeMetaInfo,
-    runtime: Arc<Runtime>,
-) -> Result<Arc<dyn MetaClient + Send + Sync>> {
-    let meta_client = MetaClientImpl::new(config, node_meta_info, runtime)?;
-    Ok(Arc::new(meta_client))
 }
